Tipos int que deveriam ser long atualizados

+++ src/lib/list.c
@@ -3,14 +3,14 @@
 
 struct llist {
   int size;
-  int * list;
+  long * list;
 };
 
 LONG_list create_list(int size) {
     if(size <= 0) return NULL;
     LONG_list l = malloc(sizeof(struct llist));
     l->size = size;
-    l->list = malloc(sizeof(int) * size);
+    l->list = malloc(sizeof(long) * size);
     return l;
 }
 
@@ -18,7 +18,7 @@ long get_list(LONG_list l, int index) {
     return l->list[index]; 
 }
 
-void set_list(LONG_list l, int index, long value) { // muda valores da lista
+void set_list(LONG_list l, int index, long value) {
     l->list[index] = value;
 }
 void free_list(LONG_list l) {
